+++ crates/interpreter/src/instructions/memory.rs
@@ -5,7 +5,7 @@ use crate::{
 };
 use core::cmp::max;
 
-pub fn mload<H: Host>(interpreter: &mut Interpreter, _host: &mut H) {
+pub fn mload<H: Host + ?Sized>(interpreter: &mut Interpreter, _host: &mut H) {
     gas!(interpreter, gas::VERYLOW);
     pop!(interpreter, index);
     let index = as_usize_or_fail!(interpreter, index);
@@ -13,7 +13,7 @@ pub fn mload<H: Host>(interpreter: &mut Interpreter, _host: &mut H) {
     push!(interpreter, interpreter.shared_memory.get_u256(index));
 }
 
-pub fn mstore<H: Host>(interpreter: &mut Interpreter, _host: &mut H) {
+pub fn mstore<H: Host + ?Sized>(interpreter: &mut Interpreter, _host: &mut H) {
     gas!(interpreter, gas::VERYLOW);
     pop!(interpreter, index, value);
     let index = as_usize_or_fail!(interpreter, index);
@@ -21,7 +21,7 @@ pub fn mstore<H: Host>(interpreter: &mut Interpreter, _host: &mut H) {
     interpreter.shared_memory.set_u256(index, value);
 }
 
-pub fn mstore8<H: Host>(interpreter: &mut Interpreter, _host: &mut H) {
+pub fn mstore8<H: Host + ?Sized>(interpreter: &mut Interpreter, _host: &mut H) {
     gas!(interpreter, gas::VERYLOW);
     pop!(interpreter, index, value);
     let index = as_usize_or_fail!(interpreter, index);
@@ -29,13 +29,13 @@ pub fn mstore8<H: Host>(interpreter: &mut Interpreter, _host: &mut H) {
     interpreter.shared_memory.set_byte(index, value.byte(0))
 }
 
-pub fn msize<H: Host>(interpreter: &mut Interpreter, _host: &mut H) {
+pub fn msize<H: Host + ?Sized>(interpreter: &mut Interpreter, _host: &mut H) {
     gas!(interpreter, gas::BASE);
     push!(interpreter, U256::from(interpreter.shared_memory.len()));
 }
 
 // EIP-5656: MCOPY - Memory copying instruction
-pub fn mcopy<H: Host, SPEC: Spec>(interpreter: &mut Interpreter, _host: &mut H) {
+pub fn mcopy<H: Host + ?Sized, SPEC: Spec>(interpreter: &mut Interpreter, _host: &mut H) {
     check!(interpreter, CANCUN);
     pop!(interpreter, dst, src, len);